# swagger.yaml
openapi: 3.0.3
info:
  title: FastForm SaaS API
  description: |
    API para la plataforma FastForm - automatización de Google Forms con integración de MercadoPago.
    
    Esta API permite:
    - Crear formularios de Google Forms desde archivos CSV/Excel
    - Generar formularios con inteligencia artificial
    - Gestionar créditos y pagos mediante MercadoPago
    - Administrar formularios existentes
    - Verificar transacciones y balances de créditos
    
    ## Autenticación
    La API utiliza autenticación basada en tokens de Firebase Auth y tokens de acceso de Google OAuth 2.0.
  version: 2.1.0
  contact:
    name: FastForm Support
    email: support@fastform.pro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://fastform.pro
    description: Servidor de producción
  - url: http://localhost:3000
    description: Servidor de desarrollo
security:
  - BearerAuth: []
  - GoogleOAuth: []
paths:
...
  /api/google-forms/create:
    post:
      tags:
        - Google Forms
      summary: Crear un nuevo formulario en Google Forms
      description: Crea un formulario en Google Forms con la estructura proporcionada
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - formData
                - accessToken
              properties:
                formData:
                  $ref: '#/components/schemas/FormData'
                accessToken:
                  type: string
                  description: Token de acceso de Google OAuth
            example:
              formData:
                title: "Encuesta de Satisfacción del Cliente"
                description: "Ayúdanos a mejorar nuestros servicios"
                questions:
                  - type: "texto_corto"
                    label: "¿Cuál es tu nombre?"
                    required: true
                  - type: "opcion_multiple"
                    label: "¿Cómo calificarías nuestro servicio?"
                    options: ["Excelente", "Bueno", "Regular", "Malo"]
                    required: true
              accessToken: "ya29.a0Ae4lvC0..."
      responses:
        '200':
          description: Formulario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      formId:
                        type: string
                        example: "1FAIpQLSdK9..."
                      formUrl:
                        type: string
                        example: "https://forms.gle/abc123"
                      editUrl:
                        type: string
                        example: "https://docs.google.com/forms/d/1FAIpQLSdK9.../edit"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/ai-chat/generate-form:
    post:
      tags:
        - Inteligencia Artificial
      summary: Generar formulario con IA
      description: Genera un formulario completo utilizando inteligencia artificial basada en una descripción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - userId
              properties:
                message:
                  type: string
                  description: Descripción del formulario que se desea crear
                userId:
                  type: string
                  description: ID del usuario de Firebase
            example:
              message: "Necesito un formulario para evaluar la satisfacción del cliente después de una compra"
              userId: "user123"
      responses:
        '200':
          description: Formulario generado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  form:
                    $ref: '#/components/schemas/GeneratedForm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # MercadoPago Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    GoogleOAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    FormData:
      type: object
      required:
        - title
        - questions
      properties:
        title:
          type: string
          description: Título del formulario
          example: "Encuesta de Satisfacción"
        description:
          type: string
          description: Descripción del formulario
          example: "Ayúdanos a mejorar nuestros servicios"
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    Question:
      type: object
      required:
        - type
        - label
      properties:
        type:
          type: string
          enum: ["texto_corto", "texto_largo", "opcion_multiple", "checkboxes", "dropdown", "escala_lineal", "fecha", "hora", "email", "numero", "archivo", "grid", "escala", "fecha_hora"]
          example: "opcion_multiple"
        label:
          type: string
          example: "¿Cómo calificarías nuestro servicio?"
        description:
          type: string
          example: "Selecciona una opción"
        required:
          type: boolean
          example: true
        options:
          type: array
          items:
            type: string
          example: ["Excelente", "Bueno", "Regular", "Malo"]
        validation:
          type: object
          properties:
            minLength:
              type: integer
            maxLength:
              type: integer
            pattern:
              type: string
    GeneratedForm:
      type: object
      properties:
        title:
          type: string
          example: "Formulario de Satisfacción del Cliente"
        description:
          type: string
          example: "Formulario generado con inteligencia artificial"
        questions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              label:
                type: string
              required:
                type: boolean
              options:
                type: array
                items:
                  type: string
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Forbidden:
      description: Permisos insuficientes
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    PaymentRequired:
      description: Créditos insuficientes
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
    RateLimit:
      description: Límite de tasa excedido
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...